# backend/internal/httpd/apispec/openapi.yaml
openapi: 3.0.3
info:
  description: Loopback API of the babysitter daemon. Generated from the Go types
    in internal/httpd; do not edit by hand.
  title: babysitter daemon API
  version: 1.0.0
servers:
- url: http://127.0.0.1:{port}/api/v1
  variables:
    port:
      default: "0"
paths:
  /api/v1/control/shutdown:
    post:
      operationId: shutdown
      responses:
        "202":
          description: Accepted
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpdAPIError'
          description: Forbidden
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpdAPIError'
          description: Internal Server Error
      summary: Stop the daemon. Rejected for browser origins.
...
  /api/v1/repos:
    get:
      operationId: listRepos
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpdRepoList'
          description: OK
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpdAPIError'
          description: Internal Server Error
      summary: List the watched repositories
    post:
      operationId: addRepo
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HttpdAddRepoRequest'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpdRepo'
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpdAPIError'
          description: Bad Request
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpdAPIError'
          description: Conflict
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HttpdAPIError'
          description: Internal Server Error
      summary: Watch a repository
...
components:
  schemas:
...
    HttpdAPIError:
      properties:
        error:
          $ref: '#/components/schemas/HttpdErrorBody'
      required:
      - error
      type: object
...
    HttpdAddRepoRequest:
      properties:
        fullName:
          type: string
      required:
      - fullName
      type: object
    HttpdErrorBody:
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      type: object
...
    HttpdRepo:
      properties:
        addedAt:
          format: date-time
          type: string
        fullName:
          type: string
        id:
          format: int64
          type: integer
        lastError:
          type: string
        lastSyncedAt:
          format: date-time
          nullable: true
          type: string
        name:
          type: string
        owner:
          type: string
      required:
      - id
      - owner
      - name
      - fullName
      - addedAt
      - lastError
      type: object
    HttpdRepoList:
      properties:
        repos:
          items:
            $ref: '#/components/schemas/HttpdRepo'
          nullable: true
          type: array
      required:
      - repos
      type: object
...